Rename: 파일 구조 변경 및 폴더명 수정
