+++ front/components/common/Paper/styles.ts
